requirements.txt
pytest
selene>=2.0.0b13
allure-pytest